Extract composer drag list
