+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.80",
-  "tag": "v3.4.80",
+  "version": "3.4.81",
+  "tag": "v3.4.81",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.79 <3.4.80"
+      ">=3.4.80 <3.4.81"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.79 first."
+    "message": "Update to v3.4.80 first."
   }
 }
